refactor: Moved test resources
Moved test resources from same file as tests to single resources file to prevent code duplication
