--- src/components/views/CorrespondenceForm/testResources.js
@@ -1,5 +1,0 @@
-  const handlers = {
-    onClose: jest.fn(),
-    onSubmit: jest.fn(),
-  };
-  export default handlers;
